isometric picking now works properly as long as picking distance = world height

+++ CubeProject/tactical/volume/ChunkManager.h
@@ -62,7 +62,7 @@ namespace tactical
 		inline void SetChunkLoadingRadius(int radius) { m_chunkLoadingRadius = radius; }
 		inline int GetChunkLoadingRadius() const { return m_chunkLoadingRadius; }
 
-		math::RayCastResult GetRayVoxelIntersection(math::Ray& ray, const glm::vec3& pos, float pickRadius);
+		math::RayCastResult GetRayVoxelIntersection(math::Ray& ray, const glm::vec3& pos, int pickRadius);
 
 		inline ChunkIterator begin() { return m_chunks.begin(); }
 		inline ChunkIterator end() { return m_chunks.end(); }
